Extract the ParallelStream class into its own file
To avoid cluttering the parallel doc with internals.
